renaming "card-file" module and leveraging in file writer & tests
